hw11/card_project/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
black==23.3.0
blinker==1.6.2
cards==1.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==7.1.2
colorama==0.4.6
commonmark==0.9.1
cryptography==41.0.2
decorator==5.1.1
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.2
fernet==1.0.1
Flask==2.0.3
Flask-Classful==0.14.2
Flask-Testing==0.8.1
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.7.0
pluggy==1.2.0
psycopg2==2.9.6
pyaes==1.6.1
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
requests==2.31.0
rich==10.7.0
self==2020.12.3
sqlparse==0.4.4
tinydb==4.5.1
tomli==2.0.1
typer==0.3.2
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
